refactor api
